% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Invoke the Authentication Process with Google}
\usage{
authenticate(save = T)
}
\arguments{
\item{save}{logical, should the authentication information should be saved on disk? Defaults to TRUE.}
}
\value{
data.frame containing credentials, optionally saved as .RData in the current working directory
}
\description{
This function starts the authentication process with
Google. Note that this function needs user interaction.
}
