% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_handle.R
\name{build_handle}
\alias{build_handle}
\title{Prepare Curl Request}
\usage{
build_handle(x)
}
\arguments{
\item{x}{object of class googleAdsSearch, googleAdsFields
listAccessibleCustomers}
}
\value{
curl dandle object
}
\description{
Create
}
