% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_auth.R
\name{do_auth}
\alias{do_auth}
\title{Invoke the Authentication Process with Google}
\usage{
do_auth(save = T)
}
\arguments{
\item{save}{logical denotes whether authentication information should be saved on disk. Defaults to TRUE.}
}
\value{
Dataframe with the credential information which is cached in working space
}
\description{
This function starts the authentication process with Google.
Note that this functions needs user interaction.
}
