% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credentials.R
\name{.get_credentials}
\alias{.get_credentials}
\title{Authentication of R app}
\usage{
.get_credentials()
}
\value{
Client token from Google authentication server.
}
\description{
get_credentials authenticates the R app at the
'Google' authentication server using OAUTH2 and receives the client token.
Usually you need not to run get_credentials() explicitly since the whole authentication process is managed by \code{\link{authenticate}}.
}
