% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_token.R
\name{.refresh_token}
\alias{.refresh_token}
\title{Refresh Access Token}
\usage{
.refresh_token(google_auth)
}
\arguments{
\item{google_auth}{list of credentials and access token}
}
\value{
New access token with corresponding time stamp
}
\description{
refresh_token returns a new valid access token. The access token deprecates after one hour and has to updated with the refresh token.
Usually you need not to run refresh_token() explicitly since the whole authentication process is managed by \code{\link{authenticate}}.
}
