% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth.R
\name{get_auth}
\alias{get_auth}
\title{Authentication of R app}
\usage{
get_auth()
}
\value{
Client token from Google authentication server.
}
\description{
get_auth authenticates the R app at the Google authentication server using OAUTH2 and receives the client token.
Usually you need not to run get_auth() explicitly since the whole authentication process is managed by \code{\link{do_auth}}.
}
