% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_services_constructors.R
\name{googleAdsFields}
\alias{googleAdsFields}
\title{Construct googleAdsFields Class (S3)}
\usage{
googleAdsFields(tbl, api_version = "v9")
}
\arguments{
\item{tbl}{character table argument}

\item{api_version}{Google Ads API version e.g 'v9'}
}
\value{
list; contains googleAdsFields Class information
}
\description{
Create an instance of a googleAdsFields class.
A list of Google's ads services and their arguments can be found here:
\url{https://developers.google.com/google-ads/api/rest/reference/rest/v9/googleAdsFields}.
}
