% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_token.R
\name{load_token}
\alias{load_token}
\title{Loading the Access Token}
\usage{
load_token(credlist)
}
\arguments{
\item{credlist}{list of credentials}
}
\value{
Access token with corresponding time stamp.
}
\description{
load_token loads the access token using credentials provided by \code{\link{get_auth}}. Execution of function is possible only once per authentication process.
Usually you need not to run load_token() explicitly since the whole authentication process is managed by \code{\link{do_auth}}.
}
