% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_services_constructors.R
\name{metrics}
\alias{metrics}
\title{Deprecated Function}
\usage{
metrics(...)
}
\description{
For more details on the Google's phase out of the Adwords API read our migration guide https://banboo-data.github.io/r4googleads
}
