% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_data}
\alias{query_data}
\title{Query Google Ads Data}
\usage{
query_data(account_id, mcc_id, google_auth,
  query = "SELECT campaign.name, campaign.status, segments.device,\\n                                metrics.impressions, metrics.clicks, metrics.ctr,\\n                                metrics.average_cpc, metrics.cost_micros\\n                                FROM campaign\\n                                WHERE segments.date DURING LAST_30_DAYS")
}
\arguments{
\item{account_id}{Google Ads Account ID}

\item{mcc_id}{Google Ads Client Center MCC Id}

\item{google_auth}{auth object}

\item{query}{Google Ads query}
}
\value{
Dataframe
}
\description{
Queries data from Google Ads API.
}
