% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_fields.R
\name{query_table_fields}
\alias{query_table_fields}
\title{Query Google Ads Fields Service}
\usage{
query_table_fields(account_id, mcc_id, google_auth, table = "campaign")
}
\arguments{
\item{account_id}{Google Ads Account ID}

\item{mcc_id}{Google Ads Client Center MCC Id}

\item{google_auth}{auth object}

\item{table}{Google Ads report tables}
}
\value{
List with field information of resource table
}
\description{
Queries table information from Google Ads Fields Service.
}
