% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R, R/deprecated.R,
%   R/google_services_constructors.R, R/query_google_ads.R
\name{do_auth}
\alias{do_auth}
\alias{r4googleads-deprecated}
\alias{statement}
\alias{reports}
\alias{metrics}
\alias{get_data}
\title{Deprecated Function}
\usage{
do_auth(...)

statement()

reports(...)

metrics(...)

get_data(...)
}
\arguments{
\item{...}{arguments to to be passed on to the underlying functions. (note that this functions is deprecated, use authenticate().)}
}
\value{
message with deprecation notice

message with deprecation notice

message with deprecation notice

message with deprecation notice

message with deprecation notice
}
\description{
You were looking for a deprecated function. Very likely you have been a user
of 'r4googleads' predecessor 'RAdwords' package. 'Google' phases out
its 'Adwords API' and replaces it with the 'Google Ads API'. Hence, we have created the
'r4googleads' package. For more information please read our migration guide
read our migration guide: \href{https://banboo-data.github.io/r4googleads/articles/radwords-migration-guide.html}{Migration Guide}
}
