% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_services_constructors.R,
%   R/query_google_ads.R
\name{statement}
\alias{statement}
\title{Deprecated Function}
\usage{
statement(...)

statement(...)
}
\description{
For more details on the Google's phase out of the Adwords API read our migration guide https://banboo-data.github.io/r4googleads

For more details on the Google's phase out of the Adwords API read our migration guide https://banboo-data.github.io/r4googleads
}
