% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Invoke the Authentication Process with Google}
\usage{
authenticate(save = T)
}
\arguments{
\item{save}{logical denotes whether authentication information should be saved on disk. Defaults to TRUE.}
}
\value{
Dataframe with the credential information which is cached in working space
}
\description{
This function starts the authentication process with Google.
Note that this function needs user interaction.
}
