% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{default_search}
\alias{default_search}
\title{Query Strings for Query Based Services}
\usage{
default_search()
}
\value{
character query statement
}
\description{
A built-in collection of queries, mostly fot the googleAdsSearch service: default_search(), ...
Usually these functions are not called stand alone. Often they are used when a constructor of a query based service is called.
}
