% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_google_ads}
\alias{query_google_ads}
\title{Query Google Ads Data}
\usage{
query_google_ads(mcc_id, google_auth, service)
}
\arguments{
\item{mcc_id}{Google Ads Client Center MCC Id}

\item{google_auth}{auth object}

\item{service}{googleAds service object created by a service constructor such as \code{\link{googleAdsSearch}} or \code{\link{googleAdsFields}}.}
}
\value{
Google Services return object either with raw or processed data (default).
}
\description{
Sends Google Ads Services objects to the Google Ads API.
}
