% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R
\name{query_google_ads}
\alias{query_google_ads}
\title{Query Google Ads Data}
\usage{
query_google_ads(account_id, mcc_id, google_auth,
  service = googleAdsSearch())
}
\arguments{
\item{account_id}{Google Ads Account ID}

\item{mcc_id}{Google Ads Client Center MCC Id}

\item{google_auth}{auth object}

\item{query}{Google Ads query}
}
\value{
Dataframe
}
\description{
Queries data from Google Ads API.
}
