% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_data.R, R/query_google_ads.R
\name{query_google_ads}
\alias{query_google_ads}
\title{Query Google Ads Data}
\usage{
query_google_ads(mcc_id, google_auth, service, raw_data = F)

query_google_ads(mcc_id, google_auth, service, raw_data = F)
}
\arguments{
\item{mcc_id}{Google Ads Client Center MCC Id}

\item{google_auth}{auth object}

\item{service}{Google Ads API Service Object}

\item{raw_data}{T/F returns raw data or content only}

\item{query}{Google Ads query}

\item{mcc_id}{Google Ads Client Center MCC Manager Id}

\item{google_auth}{auth object}
}
\value{
Dataframe

Dataframe
}
\description{
Queries data from Google Ads API.

Queries data from Google Ads API.
}
