% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R, R/deprecated.R,
%   R/google_services_constructors.R, R/query_google_ads.R
\name{do_auth}
\alias{do_auth}
\alias{r4googleads-deprecated}
\alias{statement}
\alias{metrics}
\alias{get_data}
\title{Deprecated Functions}
\usage{
do_auth(save = T)

statement()

metrics(...)

get_data(...)
}
\description{
You were looking for a deprecated function. Very likely you have been a user
of r4googleads predecessor RAdwords package. Google phases out
its AdwordsAPI and replaces it with the Google Ads API. Hence we have created the
r4googleads package. For more information please read our migration guide
 read our migration guide \code{\link{https://banboo-data.github.io/r4googleads}}
}
